+++ fiat-go/src/p256_64.go
@@ -318,7 +318,7 @@ func fiat_p256_mul(out1 *[4]uint64, arg1 *[4]uint64, arg2 *[4]uint64) {
   x187, x188 = bits.Sub64(x177, 0xffffffffffffffff, uint64(0x0))
   var x189 uint64
   var x190 uint64
-  x189, x190 = bits.Sub64(x179, uint64(0xffffffff), x188)
+  x189, x190 = bits.Sub64(x179, 0xffffffff, x188)
   var x191 uint64
   var x192 uint64
   x191, x192 = bits.Sub64(x181, uint64(0x0), x190)
@@ -622,7 +622,7 @@ func fiat_p256_square(out1 *[4]uint64, arg1 *[4]uint64) {
   x187, x188 = bits.Sub64(x177, 0xffffffffffffffff, uint64(0x0))
   var x189 uint64
   var x190 uint64
-  x189, x190 = bits.Sub64(x179, uint64(0xffffffff), x188)
+  x189, x190 = bits.Sub64(x179, 0xffffffff, x188)
   var x191 uint64
   var x192 uint64
   x191, x192 = bits.Sub64(x181, uint64(0x0), x190)
@@ -679,7 +679,7 @@ func fiat_p256_add(out1 *[4]uint64, arg1 *[4]uint64, arg2 *[4]uint64) {
   x9, x10 = bits.Sub64(x1, 0xffffffffffffffff, uint64(0x0))
   var x11 uint64
   var x12 uint64
-  x11, x12 = bits.Sub64(x3, uint64(0xffffffff), x10)
+  x11, x12 = bits.Sub64(x3, 0xffffffff, x10)
   var x13 uint64
   var x14 uint64
   x13, x14 = bits.Sub64(x5, uint64(0x0), x12)
@@ -738,7 +738,7 @@ func fiat_p256_sub(out1 *[4]uint64, arg1 *[4]uint64, arg2 *[4]uint64) {
   x10, x11 = bits.Add64(x1, (x9 & 0xffffffffffffffff), 0x0)
   var x12 uint64
   var x13 uint64
-  x12, x13 = bits.Add64(x3, (x9 & uint64(0xffffffff)), x11)
+  x12, x13 = bits.Add64(x3, (x9 & 0xffffffff), x11)
   var x14 uint64
   var x15 uint64
   x14, x15 = bits.Add64(x5, uint64(0x0), x13)
@@ -784,7 +784,7 @@ func fiat_p256_opp(out1 *[4]uint64, arg1 *[4]uint64) {
   x10, x11 = bits.Add64(x1, (x9 & 0xffffffffffffffff), 0x0)
   var x12 uint64
   var x13 uint64
-  x12, x13 = bits.Add64(x3, (x9 & uint64(0xffffffff)), x11)
+  x12, x13 = bits.Add64(x3, (x9 & 0xffffffff), x11)
   var x14 uint64
   var x15 uint64
   x14, x15 = bits.Add64(x5, uint64(0x0), x13)
@@ -946,7 +946,7 @@ func fiat_p256_from_montgomery(out1 *[4]uint64, arg1 *[4]uint64) {
   x98, x99 = bits.Sub64(x84, 0xffffffffffffffff, uint64(0x0))
   var x100 uint64
   var x101 uint64
-  x100, x101 = bits.Sub64(x88, uint64(0xffffffff), x99)
+  x100, x101 = bits.Sub64(x88, 0xffffffff, x99)
   var x102 uint64
   var x103 uint64
   x102, x103 = bits.Sub64(x94, uint64(0x0), x101)
@@ -1034,64 +1034,64 @@ func fiat_p256_to_bytes(out1 *[32]uint64, arg1 *[4]uint64) {
   var x3 uint64 = (arg1[1])
   var x4 uint64 = (arg1[0])
   var x5 uint64 = (x4 >> 8)
-  var x6 uint64 = (x4 & uint64(0xff))
+  var x6 uint64 = (x4 & 0xff)
   var x7 uint64 = (x5 >> 8)
-  var x8 uint64 = (x5 & uint64(0xff))
+  var x8 uint64 = (x5 & 0xff)
   var x9 uint64 = (x7 >> 8)
-  var x10 uint64 = (x7 & uint64(0xff))
+  var x10 uint64 = (x7 & 0xff)
   var x11 uint64 = (x9 >> 8)
-  var x12 uint64 = (x9 & uint64(0xff))
+  var x12 uint64 = (x9 & 0xff)
   var x13 uint64 = (x11 >> 8)
-  var x14 uint64 = (x11 & uint64(0xff))
+  var x14 uint64 = (x11 & 0xff)
   var x15 uint64 = (x13 >> 8)
-  var x16 uint64 = (x13 & uint64(0xff))
+  var x16 uint64 = (x13 & 0xff)
   var x17 uint64 = (x15 >> 8)
-  var x18 uint64 = (x15 & uint64(0xff))
-  var x19 uint64 = (x17 & uint64(0xff))
+  var x18 uint64 = (x15 & 0xff)
+  var x19 uint64 = (x17 & 0xff)
   var x20 uint64 = (x3 >> 8)
-  var x21 uint64 = (x3 & uint64(0xff))
+  var x21 uint64 = (x3 & 0xff)
   var x22 uint64 = (x20 >> 8)
-  var x23 uint64 = (x20 & uint64(0xff))
+  var x23 uint64 = (x20 & 0xff)
   var x24 uint64 = (x22 >> 8)
-  var x25 uint64 = (x22 & uint64(0xff))
+  var x25 uint64 = (x22 & 0xff)
   var x26 uint64 = (x24 >> 8)
-  var x27 uint64 = (x24 & uint64(0xff))
+  var x27 uint64 = (x24 & 0xff)
   var x28 uint64 = (x26 >> 8)
-  var x29 uint64 = (x26 & uint64(0xff))
+  var x29 uint64 = (x26 & 0xff)
   var x30 uint64 = (x28 >> 8)
-  var x31 uint64 = (x28 & uint64(0xff))
+  var x31 uint64 = (x28 & 0xff)
   var x32 uint64 = (x30 >> 8)
-  var x33 uint64 = (x30 & uint64(0xff))
-  var x34 uint64 = (x32 & uint64(0xff))
+  var x33 uint64 = (x30 & 0xff)
+  var x34 uint64 = (x32 & 0xff)
   var x35 uint64 = (x2 >> 8)
-  var x36 uint64 = (x2 & uint64(0xff))
+  var x36 uint64 = (x2 & 0xff)
   var x37 uint64 = (x35 >> 8)
-  var x38 uint64 = (x35 & uint64(0xff))
+  var x38 uint64 = (x35 & 0xff)
   var x39 uint64 = (x37 >> 8)
-  var x40 uint64 = (x37 & uint64(0xff))
+  var x40 uint64 = (x37 & 0xff)
   var x41 uint64 = (x39 >> 8)
-  var x42 uint64 = (x39 & uint64(0xff))
+  var x42 uint64 = (x39 & 0xff)
   var x43 uint64 = (x41 >> 8)
-  var x44 uint64 = (x41 & uint64(0xff))
+  var x44 uint64 = (x41 & 0xff)
   var x45 uint64 = (x43 >> 8)
-  var x46 uint64 = (x43 & uint64(0xff))
+  var x46 uint64 = (x43 & 0xff)
   var x47 uint64 = (x45 >> 8)
-  var x48 uint64 = (x45 & uint64(0xff))
-  var x49 uint64 = (x47 & uint64(0xff))
+  var x48 uint64 = (x45 & 0xff)
+  var x49 uint64 = (x47 & 0xff)
   var x50 uint64 = (x1 >> 8)
-  var x51 uint64 = (x1 & uint64(0xff))
+  var x51 uint64 = (x1 & 0xff)
   var x52 uint64 = (x50 >> 8)
-  var x53 uint64 = (x50 & uint64(0xff))
+  var x53 uint64 = (x50 & 0xff)
   var x54 uint64 = (x52 >> 8)
-  var x55 uint64 = (x52 & uint64(0xff))
+  var x55 uint64 = (x52 & 0xff)
   var x56 uint64 = (x54 >> 8)
-  var x57 uint64 = (x54 & uint64(0xff))
+  var x57 uint64 = (x54 & 0xff)
   var x58 uint64 = (x56 >> 8)
-  var x59 uint64 = (x56 & uint64(0xff))
+  var x59 uint64 = (x56 & 0xff)
   var x60 uint64 = (x58 >> 8)
-  var x61 uint64 = (x58 & uint64(0xff))
+  var x61 uint64 = (x58 & 0xff)
   var x62 uint64 = (x60 >> 8)
-  var x63 uint64 = (x60 & uint64(0xff))
+  var x63 uint64 = (x60 & 0xff)
   out1[0] = x6
   out1[1] = x8
   out1[2] = x10
